requirements.txt
uvicorn==0.23.2
pymongo==4.4.1
fastapi==0.101.1
pydantic==2.1.1
pandas==2.0.3
python-multipart==0.0.6
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
